Changes in GameState and UIText and deleted unnecessary classes and directiories
Added more text to the GameState class and added getter and setter to UIText so we are able to set the color of the text. After a group discussion we have decided that a splash screen is not necessary and for now we will delete the classes and directory related to it.